Add dense array implementation
